ai-service/config.py
import os
from dotenv import load_dotenv

load_dotenv()


class Settings:
    OLLAMA_BASE_URL: str = os.getenv("OLLAMA_BASE_URL", "http://localhost:11434")
    OLLAMA_MODEL: str = os.getenv("OLLAMA_MODEL", "llama3.2")
    MONGODB_URI: str = os.getenv("MONGODB_URI", "mongodb://localhost:27017/interviewprepai")
    ENVIRONMENT: str = os.getenv("ENVIRONMENT", "development")


settings = Settings()
